% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunkSnowFor.R
\name{chunkSnowFor}
\alias{chunkSnowFor}
\title{chunkSnowFor a wrapper for snowFor(), but split by chunk_size first.}
\usage{
chunkSnowFor(x, FUN, chunk_size, deley_sec = 0, ...)
}
\arguments{
\item{x}{list to loop}

\item{FUN}{the function for mapping}

\item{chunk_size}{size of *every* chunk.}

\item{deley_sec}{sleep for some time between chunks (in seconds)}

\item{...}{params here will pass to snowFor()}
}
\value{
result of snowFor()
}
\description{
chunkSnowFor a wrapper for snowFor(), but split by chunk_size first.
}
