% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txtProgressBarETA.R
\name{formatTime}
\alias{formatTime}
\title{A progress bar with estimated time to completion.}
\usage{
formatTime(seconds)
}
\description{
(rom package "pbmcapply" of Kevin kuang)
}
\details{
This is an extended version of the \code{txtProgressBar}
function from the \code{utils} package. Please refer to that for
documentation (\code{help(utils::txtProgressBar)}). The original
\code{utils::setTxtProgressBar} can be used to update the bar. Use
\code{library(pbarETA)} to override \code{utils::setTxtProgressBar}
with \code{pbarETA::setTxtProgressBar}. Use
\code{help(setTxtProgressBar, "utils")} to get help about the
original function.
}
\author{
Francesco Napolitano \email{franapoli@gmail.com}
}
