% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadError.R
\name{getErrors}
\alias{getErrors}
\title{getErrors filter error ThreadError from result of snowFor()}
\usage{
getErrors(x)
}
\arguments{
\item{x}{return of snwoFor()}
}
\description{
getErrors filter error ThreadError from result of snowFor()
}
