% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowFor.R
\name{snowFor}
\alias{snowFor}
\title{snowFor 利用snow和foreach的多线程map函数}
\usage{
snowFor(x, FUN, pre_fun = NULL, cores = parallel::detectCores(),
  env = globalenv())
}
\arguments{
\item{x}{一个list，传给foreach的第一个参数}

\item{FUN}{要运行的函数}

\item{pre_fun}{预备函数，用于初始化每个节点}

\item{cores}{现成数量，默认为全部逻辑核心数}

\item{env}{存放cl对象的环境，默认为globalenv()}
}
\description{
用法等同于lapply。但出错不会中断，会返回一个ThreadError对象。
}
\examples{
go_fun = function(x){
  Sys.sleep(1)
  x
 }
a = snowFor(1:10,go_fun,cores = 2)


}
