% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadError.R
\name{ThreadError}
\alias{ThreadError}
\title{ThreadError Error object for a thread}
\usage{
ThreadError(e, param)
}
\arguments{
\item{e}{error}

\item{param}{the loop parameter when the error happen}
}
\description{
ThreadError Error object for a thread
}
